+++ lib/dalli/client.rb
@@ -17,7 +17,6 @@ module Dalli
     # Options:
     # - :namespace - prepend each key with this value to provide simple namespacing.
     # - :failover - if a server is down, look for and store values on another server in the ring.  Default: true.
-    # - :nonascii - allow the use of nonascii key names.  Default: false.
     # - :threadsafe - ensure that only one thread is actively using a socket at a time. Default: true.
     # - :expires_in - default TTL in seconds if you do not pass TTL as a parameter to an individual operation, defaults to 0 or forever
     # - :compress - defaults to false, if true Dalli will compress values larger than 100 bytes before