+++ History.md
@@ -57,6 +57,8 @@ Notes:
     require 'rack/session/dalli'
     use Rack::Session::Dalli, :memcache_server => 'localhost:11211', :compression => true
 
+- Add support for non-ascii keys
+
 1.1.3
 =======